requirements.txt
alembic==1.4.2
beautifulsoup4==4.9.0
blinker==1.4
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.5.1
Mako==1.1.2
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-docx==0.8.10
python-editor==1.0.4
requests==2.23.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.16
urllib3==1.25.9
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1